move integration test app into own folder
