+++ packages/electron/src/integration-test/tsconfig.json
@@ -8,6 +8,6 @@
     "outDir": "../../dist",
     "sourceMap": true
   },
-  "include": ["./index.ts", "../renderer/ElectronPreload.ts"],
+  "include": ["./test-app/index.ts", "../renderer/ElectronPreload.ts"],
   "exclude": ["node_modules"]
 }
